+++ test/messageRoute.js
@@ -5,7 +5,9 @@ process.env.NODE_ENV = 'test';
 let chai = require('chai');
 let chaiHttp = require('chai-http');
 
-let server = require('../index').app;
+let theapp = require('../index');
+theapp.logger.setLog("debug");
+let server = theapp.app;
 
 // async function start() {
 //     await require('../index').startServer();
@@ -31,30 +33,73 @@ var assert = chai.assert;
 
 chai.use(chaiHttp);
 
-before(async () => {
-  require('../index').logger.setLog("ERROR");
-  await require('../index').startAMQP();
-  var startServer = await observer.startServer({ rabbitmq_uri: process.env.RABBITMQ_URI });
-  console.log("startServer before", startServer)
+console.log("\n\n****************************** WARNING ********************************")
+console.log("***********************************************************************")
+console.log("******************** ONLY MONGODB AND RABBITMQ ON *********************")
+console.log("***********************************************************************")
+
+const user1 = {
+  id: '5f09983d20f76b0019af7190',
+  fullname: 'Andrea Leo',
+  firstname: 'Andrea',
+  lastname: 'Leo',
+  token: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJqdGkiOiJmYmY2ODczMy1lZjRjLTQ2YmItOGU3ZS0wMWMyNWZkMDdhZGIiLCJzdWIiOiI1ZjA5OTgzZDIwZjc2YjAwMTlhZjcxOTAiLCJzY29wZSI6WyJyYWJiaXRtcS5yZWFkOiovKi9hcHBzLnRpbGVjaGF0LnVzZXJzLjVmMDk5ODNkMjBmNzZiMDAxOWFmNzE5MC4qIiwicmFiYml0bXEud3JpdGU6Ki8qL2FwcHMudGlsZWNoYXQudXNlcnMuNWYwOTk4M2QyMGY3NmIwMDE5YWY3MTkwLioiLCJyYWJiaXRtcS5jb25maWd1cmU6Ki8qLyoiXSwiY2xpZW50X2lkIjoiNWYwOTk4M2QyMGY3NmIwMDE5YWY3MTkwIiwiY2lkIjoiNWYwOTk4M2QyMGY3NmIwMDE5YWY3MTkwIiwiYXpwIjoiNWYwOTk4M2QyMGY3NmIwMDE5YWY3MTkwIiwidXNlcl9pZCI6IjVmMDk5ODNkMjBmNzZiMDAxOWFmNzE5MCIsImFwcF9pZCI6InRpbGVjaGF0IiwiaWF0IjoxNjE1Mjg2MzIxLCJleHAiOjE5MjYzMjYzMjEsImF1ZCI6WyJyYWJiaXRtcSIsIjVmMDk5ODNkMjBmNzZiMDAxOWFmNzE5MCJdLCJraWQiOiJ0aWxlZGVzay1rZXkiLCJ0aWxlZGVza19hcGlfcm9sZXMiOiJ1c2VyIn0.3Gt5_rT1lwmvV0wEoFYMedUFt25UIVbF-Qt3ufjPjQ4'
+}
+// RABBIT USER (nico.lanzo@frontiere21.it) TOKEN:
+const user2 = {
+  id: '82004a48ed067c0012dd32dd',
+  fullname: 'Nico Lanzo',
+  firstname: 'Nico',
+  lastname: 'Lanzo',
+  token: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJqdGkiOiI2ZmY0YTk0Yy0zMjBlLTRhYmItYTExZi00MTM3OTlmYmU2YzkiLCJzdWIiOiI4MjAwNGE0OGVkMDY3YzAwMTJkZDMyZGQiLCJzY29wZSI6WyJyYWJiaXRtcS5yZWFkOiovKi9hcHBzLnRpbGVjaGF0LnVzZXJzLjgyMDA0YTQ4ZWQwNjdjMDAxMmRkMzJkZC4qIiwicmFiYml0bXEud3JpdGU6Ki8qL2FwcHMudGlsZWNoYXQudXNlcnMuODIwMDRhNDhlZDA2N2MwMDEyZGQzMmRkLioiLCJyYWJiaXRtcS5jb25maWd1cmU6Ki8qLyoiXSwiY2xpZW50X2lkIjoiODIwMDRhNDhlZDA2N2MwMDEyZGQzMmRkIiwiY2lkIjoiODIwMDRhNDhlZDA2N2MwMDEyZGQzMmRkIiwiYXpwIjoiODIwMDRhNDhlZDA2N2MwMDEyZGQzMmRkIiwidXNlcl9pZCI6IjgyMDA0YTQ4ZWQwNjdjMDAxMmRkMzJkZCIsImFwcF9pZCI6InRpbGVjaGF0IiwiaWF0IjoxNjE1Mjg2MzIxLCJleHAiOjE5MjYzMjYzMjEsImF1ZCI6WyJyYWJiaXRtcSIsIjgyMDA0YTQ4ZWQwNjdjMDAxMmRkMzJkZCJdLCJraWQiOiJ0aWxlZGVzay1rZXkiLCJ0aWxlZGVza19hcGlfcm9sZXMiOiJ1c2VyIn0.3iA2cw7YpiKhqva3E8US9xx-mHS6t14ZuvA4nWMhEio'
+}
+
+before( () => {
+  return new Promise( async (resolve, reject) => {
+    require('../index').logger.setLog("ERROR");
+    await require('../index').startAMQP({
+      CONTACTS_LOOKUP_ENDPOINT: "http://localhost:3002/contacts"
+    });
+    observer.setWebHookEnabled(true);
+    observer.logger.setLog("error");
+    var startServer = await observer.startServer(
+    {
+      rabbitmq_uri: process.env.RABBITMQ_URI,
+      mongodb_uri: process.env.MONGODB_URI
+    });
+    console.log("startServer before", startServer)
+    console.log("Starting contacts server...");
+    var contactsServer = express();
+    contactsServer.use(bodyParser.json());
+    contactsServer.get("/contacts/:userid", (req, res) => {
+      const contacts = {
+        "5f09983d20f76b0019af7190": user1,
+        "82004a48ed067c0012dd32dd": user2
+      }
+      const userid = req.params.userid
+      console.log("/chat21/contacts/:userid", userid)
+      let contact = contacts[userid];
+      console.log("user:", contact)
+      if (!contact) {
+          contact = {
+              _id: userid,
+              "fullname": userid + " NOT FOUND"
+          };
+      }
+      contact.uid = contact.id;
+      contact.description = 'id:' + contact.uid;
+      res.status(200).send(contact);
+      res.end();
+    });
+    var listener = contactsServer.listen(3002, '0.0.0.0',  () => {
+      console.log('contactsServer started.', listener.address());
+      resolve();
+    });
+  })
+  
 });
 
 describe('MessageRoute', () => {
-  const user1 = {
-    id: '5f09983d20f76b0019af7190',
-    fullname: 'Andrea Leo',
-    // firstname: 'Andrea',
-    // lastname: 'Leo',
-    token: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJqdGkiOiJmYmY2ODczMy1lZjRjLTQ2YmItOGU3ZS0wMWMyNWZkMDdhZGIiLCJzdWIiOiI1ZjA5OTgzZDIwZjc2YjAwMTlhZjcxOTAiLCJzY29wZSI6WyJyYWJiaXRtcS5yZWFkOiovKi9hcHBzLnRpbGVjaGF0LnVzZXJzLjVmMDk5ODNkMjBmNzZiMDAxOWFmNzE5MC4qIiwicmFiYml0bXEud3JpdGU6Ki8qL2FwcHMudGlsZWNoYXQudXNlcnMuNWYwOTk4M2QyMGY3NmIwMDE5YWY3MTkwLioiLCJyYWJiaXRtcS5jb25maWd1cmU6Ki8qLyoiXSwiY2xpZW50X2lkIjoiNWYwOTk4M2QyMGY3NmIwMDE5YWY3MTkwIiwiY2lkIjoiNWYwOTk4M2QyMGY3NmIwMDE5YWY3MTkwIiwiYXpwIjoiNWYwOTk4M2QyMGY3NmIwMDE5YWY3MTkwIiwidXNlcl9pZCI6IjVmMDk5ODNkMjBmNzZiMDAxOWFmNzE5MCIsImFwcF9pZCI6InRpbGVjaGF0IiwiaWF0IjoxNjE1Mjg2MzIxLCJleHAiOjE5MjYzMjYzMjEsImF1ZCI6WyJyYWJiaXRtcSIsIjVmMDk5ODNkMjBmNzZiMDAxOWFmNzE5MCJdLCJraWQiOiJ0aWxlZGVzay1rZXkiLCJ0aWxlZGVza19hcGlfcm9sZXMiOiJ1c2VyIn0.3Gt5_rT1lwmvV0wEoFYMedUFt25UIVbF-Qt3ufjPjQ4'
-  }
-  // RABBIT USER (nico.lanzo@frontiere21.it) TOKEN:
-  const user2 = {
-    id: '82004a48ed067c0012dd32dd',
-    fullname: 'Nico Lanzo',
-    //  firstname: 'Nico',
-    //  lastname: 'Lanzo',
-    token: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJqdGkiOiI2ZmY0YTk0Yy0zMjBlLTRhYmItYTExZi00MTM3OTlmYmU2YzkiLCJzdWIiOiI4MjAwNGE0OGVkMDY3YzAwMTJkZDMyZGQiLCJzY29wZSI6WyJyYWJiaXRtcS5yZWFkOiovKi9hcHBzLnRpbGVjaGF0LnVzZXJzLjgyMDA0YTQ4ZWQwNjdjMDAxMmRkMzJkZC4qIiwicmFiYml0bXEud3JpdGU6Ki8qL2FwcHMudGlsZWNoYXQudXNlcnMuODIwMDRhNDhlZDA2N2MwMDEyZGQzMmRkLioiLCJyYWJiaXRtcS5jb25maWd1cmU6Ki8qLyoiXSwiY2xpZW50X2lkIjoiODIwMDRhNDhlZDA2N2MwMDEyZGQzMmRkIiwiY2lkIjoiODIwMDRhNDhlZDA2N2MwMDEyZGQzMmRkIiwiYXpwIjoiODIwMDRhNDhlZDA2N2MwMDEyZGQzMmRkIiwidXNlcl9pZCI6IjgyMDA0YTQ4ZWQwNjdjMDAxMmRkMzJkZCIsImFwcF9pZCI6InRpbGVjaGF0IiwiaWF0IjoxNjE1Mjg2MzIxLCJleHAiOjE5MjYzMjYzMjEsImF1ZCI6WyJyYWJiaXRtcSIsIjgyMDA0YTQ4ZWQwNjdjMDAxMmRkMzJkZCJdLCJraWQiOiJ0aWxlZGVzay1rZXkiLCJ0aWxlZGVza19hcGlfcm9sZXMiOiJ1c2VyIn0.3iA2cw7YpiKhqva3E8US9xx-mHS6t14ZuvA4nWMhEio'
-  }
-
   describe('/sendDirect', () => {
     // mocha test/messageRoute.js  --grep 'sendDirectSent'      
     it('sendDirectSent', (done) => {
@@ -93,24 +138,23 @@ describe('MessageRoute', () => {
         // per channel_type direct nn partono i webhook
         .send({ sender_fullname: user1.fullname, recipient_id: user2.id, recipient_fullname: user2.fullname, text: "text-sendDirectSent" })
         .end((err, res) => {
-          console.log("err", err);
+          if (err) {
+            console.error("err", err);
+          }
           console.log("res.body", res.body);
           res.should.have.status(200);
           res.body.should.be.a('object');
           expect(res.body.success).to.equal(true);
         });
     })
-    // .timeout(20000);        
-    // mocha test/messageRoute.js  --grep 'sendDirectSent'
+    
 
     it('sendDirectDelivered', (done) => {
-      // console.log("")
       observer.setWebHookEvents("message-delivered");             //NON SERVE CREDO 
       observer.getWebhooks().setWebHookEvents("message-delivered");
       console.log("observer.getWebhooks().getWebHookEvents", observer.getWebhooks().getWebHookEvents());
       observer.getWebhooks().setWebHookEndpoints(["http://localhost:3001/"]);
-      console.log("observer.getWebhooks().getWebHookEndpoint", observer.getWebhooks().getWebHookEndpoints());
-      // observer.setWebHookEvents("message-sent");         
+      console.log("observer.getWebhooks().getWebHookEndpoint", observer.getWebhooks().getWebHookEndpoints());         
       var serverClient = express();
       serverClient.use(bodyParser.json());
       serverClient.post('/', function (req, res) {
@@ -118,15 +162,13 @@ describe('MessageRoute', () => {
         if (req.body.event_type == "message-delivered") {
           console.log('serverClient req', JSON.stringify(req.body));
           console.log("serverClient.headers", JSON.stringify(req.headers));
-          // console.log("111",req.body.data.text);
           expect(req.body.data.text).to.equal("text-sendDirectDelivered");
           expect(req.body.data.recipient).to.equal(user2.id);
           expect(req.body.data.status).to.equal(150);
           expect(req.body.data.sender).to.equal(user1.id);
           expect(req.body.data.sender_fullname).to.equal(user1.fullname);
           expect(req.body.data.recipient_fullname).to.equal(user2.fullname);
-          expect(req.body.data.channel_type).to.equal("direct");
-          // expect(req.body.data.timelineOf).to.equal("5f09983d20f76b0019af7190");           
+          expect(req.body.data.channel_type).to.equal("direct");        
           res.send({ text: "ok from webhook" });
           listener.close(function () { console.log('listener closed :('); });
           done();
@@ -156,26 +198,19 @@ describe('MessageRoute', () => {
       var group_members = {};
       group_members[user2.id] = 1;
       const SENT_MESSAGE = "text-sendGroupSent";
-
-      observer.getWebhooks().setWebHookEvents("message-sent");
+      observer.setWebHookEvents(["message-sent"]);
       console.log("observer.getWebhooks().getWebHookEvents", observer.getWebhooks().getWebHookEvents());
-
       // observer.setWebHookEndpoint("http://localhost:3001/");
-      observer.getWebhooks().setWebHookEndpoints(["http://localhost:3001/"]);
-      console.log("observer.getWebhooks().getWebHookEndpoint", observer.getWebhooks().getWebHookEndpoints());
+      observer.setWebHookEndpoints(["http://localhost:3001/"]);
+      console.log("observer.getWebhooks().getWebHookEndpoints()", observer.getWebhooks().getWebHookEndpoints());
       // observer.setWebHookEvents("message-sent");
-
       var count = 0;
       var serverClient = express();
       serverClient.use(bodyParser.json());
       serverClient.post('/', function (req, res) {
-        console.log("res.body webhook:", req.body);
-        console.log("res.body webhook:", JSON.stringify(req.body));
+        //console.log("req.body webhook:", req.body);
+        console.log("req.body webhook:", JSON.stringify(req.body));
         if (req.body.event_type == "message-sent") {
-          // console.log('serverClient req', JSON.stringify(req.body));
-          // console.log("serverClient.headers",  JSON.stringify(req.headers));
-          // console.log("111",req.body.data.text);
-          // expect(req.body.data.text).to.equal("Group created");
           expect(req.body.data.channel_type).to.equal("group");
           console.log("ok1");
           expect(req.body.data.recipient).to.equal(recipient_id);
@@ -192,15 +227,6 @@ describe('MessageRoute', () => {
             console.log("SENT_MESSAGE ok");
             count++;
           }
-          // if (req.body.data.text === "text-sendGroupSent") {
-          //   count++;
-          //   expect(req.body.data.sender).to.equal(user1.id);
-          //   console.log("ok5.0");
-          //   expect(req.body.data.sender_fullname).to.equal(user1.fullname);
-          //   console.log("ok6.0");
-          //   expect(req.body.data.text).to.equal("text-sendGroupSent");
-          //   console.log("ok7.0");
-          // }
           else if (
             req.body.data.sender === 'system' &&
             req.body.data.attributes.messagelabel.key === "MEMBER_JOINED_GROUP" &&
@@ -217,15 +243,6 @@ describe('MessageRoute', () => {
             console.log("82004a48ed067c0012dd32dd MEMBER_JOINED_GROUP ok");
             count++;
           }
-          // else if (req.body.data.text === "5f09983d20f76b0019af7190 added to group") {
-          //   count++;
-          //   expect(req.body.data.sender).to.equal("system");
-          //   console.log("ok5.1");
-          //   expect(req.body.data.sender_fullname).to.equal("System");   
-          //   console.log("ok6.1");
-          //   expect(req.body.data.text).to.equal("5f09983d20f76b0019af7190 added to group");
-          //   console.log("ok7.1");
-          // }
           else if (
             req.body.data.sender === 'system' &&
             req.body.data.attributes.messagelabel.key === "GROUP_CREATED" &&
@@ -234,26 +251,6 @@ describe('MessageRoute', () => {
             console.log("5f09983d20f76b0019af7190 GROUP_CREATED ok");
             count++;
           }
-          // else if (req.body.data.text === "82004a48ed067c0012dd32dd added to group") {
-          //   count++;
-          //   expect(req.body.data.sender).to.equal("system");   
-          //   console.log("ok5.2");
-          //   expect(req.body.data.sender_fullname).to.equal("System");   
-          //   console.log("ok6.2");                                                                   
-          //   expect(req.body.data.attributes.messagelabel.key).to.equal("MEMBER_JOINED_GROUP");
-          //   console.log("ok7.2");
-          //   expect(req.body.data.attributes.messagelabel.parameters.member_id).to.equal("82004a48ed067c0012dd32dd");
-          //   console.log("ok8.2");
-          // }
-          // else if (req.body.data.text === "Group created") {
-          //   count++;
-          //   expect(req.body.data.sender).to.equal("system");
-          //   console.log("ok5.3");
-          //   expect(req.body.data.sender_fullname).to.equal("System");
-          //   console.log("ok6.3");                                                                  
-          //   expect(req.body.data.text).to.equal("Group created");
-          //   console.log("ok7.3");
-          // }
           else {
             console.log("other text", req.body.data.text);
           }
@@ -280,6 +277,7 @@ describe('MessageRoute', () => {
           // per channel_type direct nn partono i webhook
           .send({ group_id: recipient_id, group_name: GROUP_NAME, group_members: group_members, attributes: { a1: "v1" } })
           .end((err, res) => {
+            console.log("group created!");
             console.log("err", err);
             console.log("res.body", res.body);
             res.should.have.status(201);
@@ -296,7 +294,7 @@ describe('MessageRoute', () => {
               .post('/api/tilechat/messages')
               .set({ "Authorization": `${user1.token}` })
               // per channel_type direct nn partono i webhook
-              .send({ sender_fullname: user1.fullname, recipient_id: recipient_id, recipient_fullname: GROUP_NAME, channel_type: "group", text: "text-sendGroupSent" })
+              .send({ sender_fullname: user1.fullname, recipient_id: recipient_id, recipient_fullname: GROUP_NAME, channel_type: "group", text: SENT_MESSAGE }) //"text-sendGroupSent"
               .end((err, res) => {
                 console.log("err", err);
                 console.log("res.body", res.body);
@@ -305,17 +303,8 @@ describe('MessageRoute', () => {
                 expect(res.body.success).to.equal(true);
               });
           });
-
       });
-
-
     });
-
-
-
-
-    // .timeout(20000);        
-
   });